+++ packages/terra-date-picker/tests/jest/DatePicker.test.jsx
@@ -54,6 +54,12 @@ it('should render a date picker with requestFocus', () => {
   expect(datePicker).toMatchSnapshot();
 });
 
+it('should render a date picker with onBlur', () => {
+  const onBlur = () => {};
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" onBlur={onBlur} utcOffset={0} />);
+  expect(datePicker).toMatchSnapshot();
+});
+
 it('should render a date picker with onChange', () => {
   const onChange = () => {};
   const datePicker = shallowWithIntl(<DatePicker name="date-input" onChange={onChange} utcOffset={0} />);
@@ -88,9 +94,8 @@ it('should render a date picker with onCalendarButtonClick', () => {
   expect(datePicker).toMatchSnapshot();
 });
 
-it('should render a date picker with onInputFocus', () => {
-  const handleOnInputFocus = () => {};
-  const datePicker = shallowWithIntl(<DatePicker name="date-input" onInputFocus={handleOnInputFocus} utcOffset={0} />);
+it('should render a date picker with onFocus', () => {
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" onFocus={() => {}} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });